chore: update comments and create react-dom monkey patch
