+++ packages/paste-theme/__tests__/themeConsumer.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {attachRootToDom} from 'test-tools/react-dom';
+import {createRoot} from 'react-dom/client';
 import {render, act} from '@testing-library/react';
 
 import {Theme} from '../src';
@@ -10,7 +10,7 @@ const ThemeConsumerExampleComponent = (): React.ReactElement => {
 
 describe('Theme.Consumer', () => {
   it('should render without crashing', (): void => {
-    const root = attachRootToDom(document.createElement('div') as HTMLElement);
+    const root = createRoot(document.createElement('div') as HTMLElement);
     act(() => {
       root.render(
         <Theme.Provider theme="default">
