+++ packages/paste-theme/__tests__/themeProvider.spec.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {attachRootToDom} from 'test-tools/react-dom';
+import {createRoot} from 'react-dom/client';
 import {render, act} from '@testing-library/react';
 
 import {Theme} from '../src';
@@ -18,7 +18,7 @@ const ThemeConsumerExampleTextColor = (): React.ReactElement => {
 
 describe('Theme.Provider', () => {
   it('should render without crashing', (): void => {
-    const root = attachRootToDom(document.createElement('div') as HTMLElement);
+    const root = createRoot(document.createElement('div') as HTMLElement);
     act(() => {
       root.render(<Theme.Provider />);
     });
