+++ tools/test/act-user-event.ts
@@ -1,3 +1,6 @@
+/*
+ *
+ */
 import userEvent from '@testing-library/real-user-event';
 import {act} from '@testing-library/react';
